refactor live assistant workflow orchestration
